requirements.txt
openai>=1.52.0
playwright>=1.47.0
python-dotenv>=1.0.1
aiohttp>=3.10.5
numpy
websockets>=12.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
pyaudio>=0.2.13
azure-cosmos>=4.5.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
